Drop internal 'modules' package
